+++ lib/widgets/pages/export_data_page/export_ride_page.dart
@@ -3,8 +3,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:weforza/generated/l10n.dart';
 import 'package:weforza/model/export/export_rides_delegate.dart';
 import 'package:weforza/model/ride.dart';
-import 'package:weforza/riverpod/file_storage_delegate_provider.dart';
-import 'package:weforza/riverpod/file_system_provider.dart';
+import 'package:weforza/riverpod/file/file_storage_delegate_provider.dart';
+import 'package:weforza/riverpod/file/file_system_provider.dart';
 import 'package:weforza/riverpod/repository/export_rides_repository_provider.dart';
 import 'package:weforza/widgets/custom/animated_circle_checkmark.dart';
 import 'package:weforza/widgets/pages/export_data_page/export_data_page.dart';
